NewlineText now allows for hyperlinks
